#10 Move function to utility collection
